refactor: split courseOffering types into files
There were too many types and interfaces in a single file. Now, the
types and interfaces used by the courseOffering wrapper function have
been split into more files.
