+++ src/api-types/courseOfferings.ts
@@ -1,106 +1,14 @@
-export enum DeliveryMethod {
-    InPerson = 'In Person',
-    DistanceEducation = 'Distance Education',
-}
-
-export enum DegreeLevel {
-    Undergraduate = 'UGRD',
-    Graduate = 'GRAD',
-}
-
-export enum Enrollment {
-    EnrollmentSection = 'e',
-    NonEnrollment = 'n',
-}
-
-export enum InstructorRole {
-    PrimaryInstructor = 'PI',
-    SecondaryInstructor = 'SI',
-}
-
-export interface RawApiInstructor {
-    profileUrl: string;
-    commonName?: string;
-    firstName: string;
-    lastName: string;
-    phone: string;
-    roleCode: InstructorRole;
-    name: string;
-    officeHours: string;
-    office: string;
-    email: string;
-}
-
-export interface ProcessedApiInstructor {
-    profileUrl: string;
-    commonName?: string;
-    firstName: string;
-    lastName: string;
-    phone: string;
-    role: InstructorRole;
-    name: string;
-    officeHours: string;
-    office: string;
-    email: string;
-}
-
-export enum Day {
-    Monday = 'Mo',
-    Tuesday = 'Tu',
-    Wednesday = 'We',
-    Thursday = 'Th',
-    Friday = 'Fr',
-    Saturday = 'Sa',
-    Sunday = 'Su',
-}
-
-export enum Campus {
-    Burnaby = 'Burnaby',
-    Surrey = 'Surrey',
-    Vancouver = 'Vancouver',
-}
-
-export enum SectionCode {
-    Lecture = 'LEC',
-    Tutorial = 'TUT',
-    Lab = 'LAB',
-    Seminar = 'SEM',
-}
-
-export interface RawApiSchedulePart {
-    roomNumber: number;
-    endDate: string;
-    campus: Campus;
-    buildingCode: string;
-    days: string;
-    sectionCode: SectionCode;
-    startTime: string;
-    isExam: boolean;
-    endTime: string;
-    startDate: string;
-}
-
-export interface ProcessedApiSchedulePart {
-    roomNumber: number;
-    endDate: string;
-    campus: Campus;
-    buildingCode: string;
-    days: Day[];
-    sectionCode: SectionCode;
-    startTime: string;
-    isExam: boolean;
-    endTime: string;
-    startDate: string;
-}
-
-export interface GradingScheme {
-    description: string;
-    weight: number;
-}
-
-export interface Textbook {
-    details: string;
-}
+import type {
+    DegreeLevel,
+    DeliveryMethod,
+    Enrollment,
+    GradingScheme,
+    Textbook,
+    RawApiInstructor,
+    RawApiSchedulePart,
+    ProcessedApiInstructor,
+    ProcessedApiSchedulePart,
+} from '@api-types';
 
 export interface ProcessedApiCourseOffering {
     title: string;
